Refactor app structure: split components, pages, and improve code organization
