+++ challenge-portal/src/pages/Challenges.tsx
@@ -1,0 +1,5 @@
+import ChallengeList from '../components/ChallengeList'
+
+export default function ChallengesPage() {
+  return <ChallengeList />
+}
