+++ challenge-portal/src/pages/Community.tsx
@@ -1,0 +1,5 @@
+import CommunityBoard from '../components/CommunityBoard'
+
+export default function CommunityPage() {
+  return <CommunityBoard />
+}
