+++ challenge-portal/src/pages/Leaderboard.tsx
@@ -1,0 +1,5 @@
+import LeaderboardJoy from '../components/LeaderboardJoy'
+
+export default function LeaderboardPage() {
+  return <LeaderboardJoy />
+}
