+++ challenge-portal/src/pages/Progress.tsx
@@ -1,0 +1,5 @@
+import ProgressPlayground from '../components/ProgressPlayground'
+
+export default function ProgressPage() {
+  return <ProgressPlayground />
+}
