+++ challenge-portal/src/pages/Teams.tsx
@@ -1,0 +1,5 @@
+import TeamStudio from '../components/TeamStudio'
+
+export default function TeamsPage() {
+  return <TeamStudio />
+}
